Feature/p2 flughafen datenmodell erstellen

+++ Code/airport-service/src/main/java/tbz/ch/airport/repository/AirportRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import tbz.ch.airport.entity.Airport;
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Long> {
+public interface AirportRepository extends JpaRepository<Airport, Integer> {
 }
